runtime: device: move to top level package
Let's move device package to runtime/pkg instead of being buried under
virtcontainers.
